[#183] refactor file structure
