+++ src/ml_service/.gitignore
@@ -6,15 +6,7 @@ coverage.xml
 .idea/
 ./**/*.iml
 .idea/*
-/src/ml_ser
-/precedents/text_bk/*
-/word_vectors/*.bin
-/outputs/test/*
-/outputs/*log
-/ml_models/*.bin
-/feature_extraction/clustering/optimization/*.txt
-/outputs/preprocess_model/*.bin
-/test/feature_extraction/preprocessing/test/*txt
-/outputs/**/*.txt
-/outputs/**/*.bin
-/word_vectors/cache.pickle
+/data/**/*.txt
+/data/**/*.bin
+/data/**/*.pickle
+server.log
