--- src/ml_service/precedents/.gitignore
@@ -1,1 +1,0 @@
-text_bk
