--- src/ml_service/test_data/clusters/-1.txt
@@ -1,29 +1,0 @@
- mody : le beau-frère de la locatrice
- les locateurs [
- jean-claude bédard, le locateur ?
- gaudreault, le locateur
- (ci-après le locateur)
-le locateur, m
- nahal et le locateur
- zobair, le locateur
- coretti, le locateur
-
-------------------------------------------
-
-AZ-51392311.txt
-AZ-51395603.txt
-AZ-51179312.txt
-AZ-51168259.txt
-AZ-51172306.txt
-AZ-51232537.txt
-AZ-51402327.txt
-AZ-51414731.txt
-AZ-51172143.txt
-AZ-51144313.txt
-AZ-51141726.txt
-AZ-51400593.txt
-AZ-51201816.txt
-AZ-51165223.txt
-AZ-51161615.txt
-AZ-51146951.txt
-AZ-51161120.txt
