--- src/ml_service/test_data/clusters/0.txt
@@ -1,261 +1,0 @@
-], terrebonne (québec) [
-] à montréal
-], à québec
- montréal, 500-02-126877-047, (2005) j
-], trois-rivières, québec, [
-
-------------------------------------------
-
-AZ-51227137.txt
-AZ-51168002.txt
-AZ-51167989.txt
-AZ-51168012.txt
-AZ-51168018.txt
-AZ-51168004.txt
-AZ-51167985.txt
-AZ-51168008.txt
-AZ-51168005.txt
-AZ-51167998.txt
-AZ-51167995.txt
-AZ-51167994.txt
-AZ-51167975.txt
-AZ-51168001.txt
-AZ-51168006.txt
-AZ-51168024.txt
-AZ-51168025.txt
-AZ-51168021.txt
-AZ-51168020.txt
-AZ-51168028.txt
-AZ-51167997.txt
-AZ-51168000.txt
-AZ-51167982.txt
-AZ-51167984.txt
-AZ-51167999.txt
-AZ-51412955.txt
-AZ-51176035.txt
-AZ-51168014.txt
-AZ-51167987.txt
-AZ-51198805.txt
-AZ-51167981.txt
-AZ-51167991.txt
-AZ-51220712.txt
-AZ-51167990.txt
-AZ-51168033.txt
-AZ-51168035.txt
-AZ-51167988.txt
-AZ-51168042.txt
-AZ-51168027.txt
-AZ-51168011.txt
-AZ-51176397.txt
-AZ-51167993.txt
-AZ-51168019.txt
-AZ-51168009.txt
-AZ-51167973.txt
-AZ-51168003.txt
-AZ-51168010.txt
-AZ-51168041.txt
-AZ-51167986.txt
-AZ-51168013.txt
-AZ-51168034.txt
-AZ-51162836.txt
-AZ-51168022.txt
-AZ-51168016.txt
-AZ-51167974.txt
-AZ-51407556.txt
-AZ-51401760.txt
-AZ-51183607.txt
-AZ-51407890.txt
-AZ-51414798.txt
-AZ-51161216.txt
-AZ-51170119.txt
-AZ-51188691.txt
-AZ-51172769.txt
-AZ-51407559.txt
-AZ-51185497.txt
-AZ-51227010.txt
-AZ-51393755.txt
-AZ-51398384.txt
-AZ-51168751.txt
-AZ-51396393.txt
-AZ-51398319.txt
-AZ-51398316.txt
-AZ-51234093.txt
-AZ-51189125.txt
-AZ-51398320.txt
-AZ-51173813.txt
-AZ-51229853.txt
-AZ-51398067.txt
-AZ-51175535.txt
-AZ-51163531.txt
-AZ-51229849.txt
-AZ-51235933.txt
-AZ-51235934.txt
-AZ-51185509.txt
-AZ-51396774.txt
-AZ-51178512.txt
-AZ-51163627.txt
-AZ-51398317.txt
-AZ-51235393.txt
-AZ-51402499.txt
-AZ-51169613.txt
-AZ-51402198.txt
-AZ-51141678.txt
-AZ-51141369.txt
-AZ-51229847.txt
-AZ-51163533.txt
-AZ-51401758.txt
-AZ-51161192.txt
-AZ-51164195.txt
-AZ-51398321.txt
-AZ-51398315.txt
-AZ-51168772.txt
-AZ-51180405.txt
-AZ-51188694.txt
-AZ-51168740.txt
-AZ-51392852.txt
-AZ-51414784.txt
-AZ-51178400.txt
-AZ-51392770.txt
-AZ-51177194.txt
-AZ-51168749.txt
-AZ-51228672.txt
-AZ-51398313.txt
-AZ-51163624.txt
-AZ-51224404.txt
-AZ-51415951.txt
-AZ-51168738.txt
-AZ-51197031.txt
-AZ-51229810.txt
-AZ-51402521.txt
-AZ-51183716.txt
-AZ-51141719.txt
-AZ-51392939.txt
-AZ-51396767.txt
-AZ-51402142.txt
-AZ-51235939.txt
-AZ-51404258.txt
-AZ-51398322.txt
-AZ-51398060.txt
-AZ-51178517.txt
-AZ-51402525.txt
-AZ-51229851.txt
-AZ-51235935.txt
-AZ-51185469.txt
-AZ-51197020.txt
-AZ-51229840.txt
-AZ-51188692.txt
-AZ-51407516.txt
-AZ-51229848.txt
-AZ-51168666.txt
-AZ-51162489.txt
-AZ-51176175.txt
-AZ-51185154.txt
-AZ-51392772.txt
-AZ-51229838.txt
-AZ-51173410.txt
-AZ-51161163.txt
-AZ-51183597.txt
-AZ-51229844.txt
-AZ-51181152.txt
-AZ-51161215.txt
-AZ-51398429.txt
-AZ-51228441.txt
-AZ-51228442.txt
-AZ-51229899.txt
-AZ-51168771.txt
-AZ-51396426.txt
-AZ-51235932.txt
-AZ-51141367.txt
-AZ-51407553.txt
-AZ-51398314.txt
-AZ-51168746.txt
-AZ-51407903.txt
-AZ-51185716.txt
-AZ-51185098.txt
-AZ-51185690.txt
-AZ-51396436.txt
-AZ-51173825.txt
-AZ-51407554.txt
-AZ-51407595.txt
-AZ-51406877.txt
-AZ-51398318.txt
-AZ-51168733.txt
-AZ-51229839.txt
-AZ-51402518.txt
-AZ-51181671.txt
-AZ-51235941.txt
-AZ-51407548.txt
-AZ-51225617.txt
-AZ-51226348.txt
-AZ-51168755.txt
-AZ-51401761.txt
-AZ-51235940.txt
-AZ-51398378.txt
-AZ-51218437.txt
-AZ-51398407.txt
-AZ-51141232.txt
-AZ-51196972.txt
-AZ-51163512.txt
-AZ-51230815.txt
-AZ-51396315.txt
-AZ-51188695.txt
-AZ-51225272.txt
-AZ-51401759.txt
-AZ-51177699.txt
-AZ-51407552.txt
-AZ-51196948.txt
-AZ-51413991.txt
-AZ-51188689.txt
-AZ-51223330.txt
-AZ-51197039.txt
-AZ-51188693.txt
-AZ-51229830.txt
-AZ-51402524.txt
-AZ-51189187.txt
-AZ-51415272.txt
-AZ-51161367.txt
-AZ-51168756.txt
-AZ-51399402.txt
-AZ-51169565.txt
-AZ-51185508.txt
-AZ-51188690.txt
-AZ-51398603.txt
-AZ-51141366.txt
-AZ-51393588.txt
-AZ-51173846.txt
-AZ-51235942.txt
-AZ-51173520.txt
-AZ-51228332.txt
-AZ-51185324.txt
-AZ-51141195.txt
-AZ-51231836.txt
-AZ-51189553.txt
-AZ-51161234.txt
-AZ-51228575.txt
-AZ-51235943.txt
-AZ-51163535.txt
-AZ-51179781.txt
-AZ-51141182.txt
-AZ-51168721.txt
-AZ-51229833.txt
-AZ-51414726.txt
-AZ-51401900.txt
-AZ-51185155.txt
-AZ-51175536.txt
-AZ-51401751.txt
-AZ-51402171.txt
-AZ-51141368.txt
-AZ-51173845.txt
-AZ-51175750.txt
-AZ-51173826.txt
-AZ-51402206.txt
-AZ-51185636.txt
-AZ-51415436.txt
-AZ-51197084.txt
-AZ-51392777.txt
-AZ-51415129.txt
-AZ-51402500.txt
-AZ-51180393.txt
-AZ-51183599.txt
-AZ-51201654.txt
-AZ-51177147.txt
