--- src/ml_service/test_data/clusters/1.txt
@@ -1,19 +1,0 @@
- lebrun et als
-dans l'affaire raymond-marie desjardins et als c
- robert laberge et als
-dans l'affaire de denise berthiaume et als
- fortin et als
-
-------------------------------------------
-
-AZ-51180913.txt
-AZ-51232097.txt
-AZ-51197085.txt
-AZ-51145020.txt
-AZ-51142167.txt
-AZ-51142166.txt
-AZ-51142165.txt
-AZ-51142169.txt
-AZ-51142168.txt
-AZ-51142164.txt
-AZ-51180913.txt
