--- src/ml_service/test_data/clusters/2.txt
@@ -1,35 +1,0 @@
-était locataire au bail
-le locataire doit-il des loyers?
-le locataire ne doit pas de loyer
-le locataire, m
-2 de son bail
- siconnelly qui est son locataire
-
-------------------------------------------
-
-AZ-51204815.txt
-AZ-51410777.txt
-AZ-51403993.txt
-AZ-51170026.txt
-AZ-51141952.txt
-AZ-51411931.txt
-AZ-51145001.txt
-AZ-51183539.txt
-AZ-51408488.txt
-AZ-51233394.txt
-AZ-51189381.txt
-AZ-51392443.txt
-AZ-51232519.txt
-AZ-51183203.txt
-AZ-51405709.txt
-AZ-51161120.txt
-AZ-51180314.txt
-AZ-51398301.txt
-AZ-51188735.txt
-AZ-51226646.txt
-AZ-51170819.txt
-AZ-51145940.txt
-AZ-51236035.txt
-AZ-51176213.txt
-AZ-51398323.txt
-AZ-51180304.txt
